Support `type data` out of the box
`type data` declarations cannot reasonably be given `SingKind` instances, as
these only make sense for things that exist at the term level. Whenever we pass
a `type data` declaration to the `singletons-th` machinery, we now always omit
generation of a `SingKind` instance.

This requires bumping the `th-desugar` commit to bring in the changes from
goldfirere/th-desugar#181, which are necessary for determining whether a locally
reified data type is actually a `type data` declaration or not.

Fixes #559.