Prefer user-written binder names in singled data type SAKS
Previously, the algorithm for computing a standalone kind signature for a
singled data type's standalone kind signature was poorly implemented, and it
would combine type variables from the original data type's standalone kind
signature alongside type variables from the user-written binders in the data
type declaration. The result was that it was possible for `singletons-th` to
single a data type and produce a standalone kind signature that was completely
ill-scoped, as seen in #567.

This patch implements a new algorithm that is closely based on GHC's
`matchUpSigWithDecl` function, which GHC uses to match up type variables from
data type SAKS with their user-written binders. While more verbose, the new
algorithm is (in my opinion) easier to understand and maintain. The new
algorithm will also make it much easier to add support for invisible binders in
data types declarations, which will arrive in GHC 9.8.

This bumps the `th-desugar` source-repository-package to bring in the changes
from goldfirere/th-desugar#189, which are required to make the `T567` test
case's scoping work out correctly.

Fixes #567.