Clean up local variable tracking in SgM
Previously, `SgEnv` (the reader environment for `SgM`, the singling monad) was
only tracking let-bound variables.  But as #571 reveals, it must track
_lambda_-bound variables as well, lest the `lookupVarE` function accidentally
look up a top-level function name instead of a local name that shadows it.

It turns out that `PrEnv` (the reader environment for `PrM`, the promotion
monad) was already doing this, so this patch mostly copies the existing `PrM`
treatment for lambda-bound variables over to `SgM`. While I was in town, I took
the liberty to clean up the naming conventions a bit and document how all of
this works in the new `Note [Tracking local variables]` in
`Data.Singletons.TH.Promote.Monad`.

Fixes #571.