Use noExactTyVars when promoting functions with local variables
Whenever we promote a function that closes over one or more local variables, we
now use `noExactTyVars` to ensure that none of the local variables reuse unique
`Name`s.  This isn't strictly necessary right now, but this will be required in
a future patch that will panic (triggering
https://gitlab.haskell.org/ghc/ghc/-/issues/11812) unless `noExactTyVars` is
used.

While I was in the neighborhood, I took the opportunity to write `Note
[Pitfalls of NameU/NameL]` with more context about why `noExactTyVars` is
necessary in the first place, and I added references to the Note in all the
places where we use it.