Regenerate CI to use Jammy Jellyfish, test GHC 9.10.1
This requires a more recent version of `haskell-ci` that supports GHC 9.10.1.
Another change that this more recent version of `haskell-ci` introduces is that
it drops support for pre-20.04 versions of Ubuntu, so we must also upgrade to a
more recent version of Ubuntu (I chose Ubuntu 22.04, or Jammy Jellyfish).

Unfortunately, `haskell-ci`'s GHCJS support no longer works with recent
versions of Ubuntu (see https://github.com/haskell-CI/haskell-ci/issues/723),
so I had to remove the GHCJS configuration from the CI. This does not mean that
`singletons` is dropping JavaScript support, however, as we still plan to
support building the `singletons` library with GHC's JavaScript backend (the
modern successor to GHCJS).

There is currently no way to set up a GHC JavaScript backend CI job using
`haskell-ci`, and the amount of setup required to manually create one is such
that I'm putting it off for now. If someone asks for it, we can add one later.