singletons-th: Adapt to GHC 9.10's lack of arity inference
GHC 9.10 no longer performs arity inference in type-level declarations (see
https://github.com/ghc-proposals/ghc-proposals/blob/master/proposals/0425-decl-invis-binders.rst),
and as it turns out, many of the promoted type families that `singletons-th`
generates would have the wrong arity. For instance, `singletons-th` would
promote this definition:

```hs
f :: Either a Bool
f = Right True
```

To this type family:

```hs
type F :: Either a Bool
type family F where
  F = Right True
```

With GHC 9.10 or later, however, GHC would conclude that `F` has arity 0, which
means that it should not bind any arguments (visible or invisible). The type
family equation for `F`, however, only works if `F` has arity 1! This is
because the type family equation needs to bind an invisible `@a` argument:

```hs
  F @a = Right @a @Bool True
```

To ensure that type families like `F` have the expected arity, `singletons-th`
now uses `TypeAbstractions` in more places to ensure that type family headers
bind an appropriate number of type variables, which makes the type families'
arities explicit. For instance, `singletons-th` now generates the following
code for `F`:

```hs
type F :: Either a Bool
type family F @a where -- Note the @a here, which gives it arity 1
  F = Right True
```

For more details on how this is implemented, see the new `Note [Generating type
families with the correct arity]` in `Data.Singletons.TH.Promote`.

A consequence of this change is that the average piece of
`singletons-th`–generated code is much more likely to require
`TypeAbstractions` than it did before. This explains why we now enable
`TypeAbstractions` in almost every module in `singletons-base`.

Fixes #566.