Support scoped type variables in class/instance declarations
Although #573 added some support for promoting/singling uses of scoped type
variables, it did not properly support scoped type variables in class or
instance declarations due to an oversight. This patch aims to correct that
oversight.

The key tricks to making this work are:

* When promoting class or instance methods, we explicitly quantify the type
  variables in the "helper" type family so that we can bind them on the
  left-hand sides of the promoted type family equations.
* In addition, we take care to only bring the _scoped_ type variables into
  scope over the right-hand sides of the promoted type family equations.

See the new `Note [Scoped type variables and class methods]` in
`Data.Singletons.TH.Promote.Monad` for the full details.

Fixes #581.