Don't defunctionalize helpers for class defaults or instance methods
`singletons-th` generates "helper" type families that contain the definitions
of class method defaults or instance methods. For example, this:

```hs
class C a where
  m :: a -> b -> a
  m x _ = x
```

Will be promoted to this:

```hs
class PC a where
  type M (x :: a) (y :: b) :: a
  type M x y = MHelperSym0 `Apply` x `Apply` y

type MHelper :: a -> b -> a
type family MHelper x y where
  MHelper x _ = x

type MHelperSym0 :: a ~> b ~> a
type MHelperSym1 :: a -> b ~> a
...
```

Generating defunctionalization symbols for `MHelper` is wasteful, however, as
we never really _need_ to partially apply `MHelper`. Instead, we can just
generate this code:

```hs
class PC a where
  type M (x :: a) (y :: b) :: a
  type M x y = MHelper x y
```

This takes advantage of the fact that when we apply `MHelper`, we always fully
apply it to all of its arguments. This means that we can avoid generating
defunctionalization symbols for helper type families altogether, which is a
nice optimization.

This patch implements that idea, fixing #608 in the process.